+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- Catch FMP erroneous dividend data with duplicate ex-dividend dates
+
+
 ## [0.49.1] - 2026-06-25
 ### Changed
 - Improved docs, especially for helpers and LSEG